requirements.txt
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langchain-core>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
